requirements.txt
# GeoSpatial Conservation AI - Core Dependencies
# =============================================
# Consolidated requirements for production system

# Essential Geospatial Libraries
# ==============================
geopandas>=0.13.0          # Core geospatial data manipulation
shapely>=2.0.0             # Geometric operations
pyproj>=3.4.0              # Coordinate reference systems
rasterio>=1.3.8            # Raster data I/O
folium>=0.14.0             # Interactive mapping

# Data Science & Analysis
# =======================
pandas>=2.0.0              # Data manipulation
numpy>=1.21.0              # Numerical computing
scipy>=1.9.0               # Scientific computing
scikit-learn>=1.1.0        # Machine learning

# Deep Learning & AI
# ==================
tensorflow>=2.13.0         # Deep learning framework

# Web Development & APIs
# ======================
requests>=2.28.0           # HTTP client
aiohttp>=3.8.0             # Async HTTP client
python-dotenv>=1.0.0       # Environment variable management

# Visualization
# =============
matplotlib>=3.6.0          # Static plotting
plotly>=5.11.0             # Interactive plotting
seaborn>=0.11.0            # Statistical visualization

# Interactive Applications
# ========================
streamlit>=1.28.0          # Web app framework
jupyter>=1.0.0             # Jupyter notebooks
ipykernel>=6.0.0           # Jupyter kernel

# Spatial Analysis & Statistics
# =============================
libpysal>=4.7.0            # Spatial analysis library
esda>=2.4.0                # Exploratory spatial data analysis
mapclassify>=2.4.0         # Spatial classification
pykrige>=1.6.0             # Kriging interpolation

# Network & Optimization
# ======================
networkx>=2.8.0            # Network analysis
pymoo>=0.6.0               # Multi-objective optimization

# Additional Utilities
# ===================
contextily>=1.3.0          # Basemap providers
cartopy>=0.21.0            # Cartographic projections
statsmodels>=0.14.0       # Statistical modeling
xarray>=2022.12.0          # N-dimensional arrays

# Optional: Advanced Features
# ===========================
rasterstats>=0.17.0        # Raster statistics
earthpy>=0.9.0             # Earth analytics utilities
